Remove pako dependency and move .git directory creation logic to git.ts
